Изменены типы полей Pressure и Visibility на float64 в структуре Weather; добавлен новый файл openmeteo.go с реализацией API для получения данных о погоде.

+++ internal/domain/models.go
@@ -10,8 +10,8 @@ type Today struct {
 	WindSpeed     float64
 	WindDirection uint
 	Humidity      uint
-	Pressure      uint
-	Visibility    uint
+	Pressure      float64
+	Visibility    float64
 	Precipitation float64
 	UpdatedAt     time.Time
 }
@@ -32,3 +32,5 @@ type DailyEntry struct {
 	Conditions               string
 	Date                     time.Time
 }
+
+// https://api.open-meteo.com/v1/forecast?latitude=52.52&longitude=13.41&daily=temperature_2m_min,temperature_2m_max,precipitation_probability_max,wind_speed_10m_max,weather_code&forecast_days=1
